refactor : 형변환 버그 수정

+++ src/main/java/blackjack/domain/participants/betting/Profit.java
@@ -17,7 +17,7 @@ public class Profit {
     }
 
     public Profit multiple(double ratio) {
-        return new Profit((int) Math.floor(profit * ratio), isPositive);
+        return new Profit((long) Math.floor(profit * ratio), isPositive);
     }
 
     public Profit inverse() {
